+++ GameProjectFinal/GameProjectFinal/Direct3D.cpp
@@ -121,7 +121,7 @@ bool Direct3D::Initialize(HWND hwnd)
 	m_videoCardMemory = (int)(adapterDesc.DedicatedVideoMemory / 1024 / 1024);
 
 	// Convert the name of the video card to a character array and store it.
-	error = wcstombs_s((size_t*)&stringLength, m_videoCardDescription, 128, adapterDesc.Description, 128);
+	error = wcstombs_s((unsigned int*)&stringLength, m_videoCardDescription, 128, adapterDesc.Description, 128);
 	if (error != 0)
 	{
 		return false;
